Build locales from moment-locales
The plan now is that users will import the locales they want from
`fecha-locales/<locale-name>`. For building, we run `npm run
build-locales`, which convert all the moment-locales in
`/locale-builder/moment-locales` and write them to files with same name
in `/`. However, all the locales don't get converted smoothely. For the
ones that fail, the script will show an error and write a file
~<localename>~.js, and we will need to fix the issue by hand.